+++ .gitignore
@@ -3,6 +3,8 @@
 __pycache__/
 .pytest_cache/
 
+# Log file
+*.log
 
 # virtual environment
 .venv*/
