+++ server/logger_default_settings.json
@@ -1,0 +1,31 @@
+{
+    "version": 1,
+    "formatters": {
+        "default": {
+            "format": "%(asctime)s %(name)s - %(levelname)s : %(message)s"
+        },
+        "to_file": {
+            "format": "[%(asctime)s] %(levelname)s in %(module)s: %(message)s"
+        }
+    },
+    "handlers": {
+        "console": {
+            "class": "logging.StreamHandler",
+            "level": "INFO",
+            "formatter": "default",
+            "stream": "ext://sys.stdout"
+        },
+        "file": {
+            "class": "logging.handlers.RotatingFileHandler",
+            "formatter": "to_file",
+            "filename": "server.log",
+            "maxBytes": 500,
+            "backupCount": 3,
+            "level": "DEBUG"
+        }
+    },
+    "root": {
+        "level": "DEBUG",
+        "handlers": ["console", "file"]
+    }
+}
